+++ airops_weather/examples/hems_integration/fxmanifest.lua
@@ -1,0 +1,7 @@
+fx_version 'cerulean'
+game 'gta5'
+lua54 'yes'
+
+server_script 'server.lua'
+
+dependency 'airops_weather'